feat: built-in indicatif progress handler behind default-off feature
Add an `indicatif` cargo feature (default off, native only) that ships a
ready-made `IndicatifProgress` terminal progress handler in the library,
moved in from the hfrs CLI's `CliProgressHandler`.

The 1.0 `ProgressHandler` trait replaced 0.x's free `with_progress(true)`
progress bar, so migrated consumers lost download progress UX unless they
hand-rolled an indicatif handler. This ships that handler in-library so
any consumer gets it back with one feature flag and one line.

- hf-hub: optional `indicatif` dep declared under the non-wasm target table
  so enabling the feature on wasm is a no-op; module cfg-gated on
  `not(target_family = "wasm")`.
- hfrs: enables the feature and uses `IndicatifProgress::with_multi`; its
  own handler implementation is deleted (the CLI-only `MultiProgressWriter`
  log adapter and env helper stay). CLI output is unchanged.
- Adds `examples/progress_indicatif.rs` and updates the AGENTS.md feature
  table and lib.rs feature docs.
