+++ examples/Cargo.toml
@@ -14,6 +14,7 @@ serde_json = "1"
 
 [features]
 blocking = ["hf-hub/blocking"]
+indicatif = ["hf-hub/indicatif"]
 
 [[example]]
 name = "repository"
@@ -54,6 +55,11 @@ path = "progress.rs"
 [[example]]
 name = "progress_logger"
 path = "progress_logger.rs"
+
+[[example]]
+name = "progress_indicatif"
+path = "progress_indicatif.rs"
+required-features = ["indicatif"]
 
 [[example]]
 name = "buckets"
