+++ hfrs/src/commands/buckets/cp.rs
@@ -5,10 +5,9 @@ use anyhow::Result;
 use clap::Args as ClapArgs;
 use hf_hub::HFClient;
 use hf_hub::buckets::{BucketCopySourceType, BucketDownload, BucketUpload};
-use hf_hub::progress::Progress;
+use hf_hub::progress::{IndicatifProgress, Progress};
 
 use crate::output::CommandResult;
-use crate::progress::CliProgressHandler;
 
 #[derive(ClapArgs)]
 pub struct Args {
@@ -50,7 +49,7 @@ pub async fn execute(client: &HFClient, args: Args, multi: Option<indicatif::Mul
     let handler: Option<Progress> = if args.quiet {
         None
     } else {
-        multi.map(|multi| CliProgressHandler::new(multi).into())
+        multi.map(|multi| IndicatifProgress::with_multi(multi).into())
     };
 
     let src_is_stdin = args.src == "-";
